Default STAC collection

Endpoint and collection ID of the default STAC collection where to access digital elevation model (DEM) data. This is the global Copernicus DEM 30 dataset hosted on AWS, as listed in the EarthSearch STAC API endpoint. Note that AWS credentials need to be set up in order to access the data (not the catalog). References:

EarthSearch STAC API

Copernicus DEM

AWS Copernicus DEM datasets

Data license

Usage

default_stac_dem

Format

An object of class list of length 2.