Delineate a corridor around a river.

Delineate a corridor around a river.

Usage

delineate(
  city_name,
  river_name,
  crs = NULL,
  network_buffer = NULL,
  buildings_buffer = NULL,
  corridor_init = "valley",
  dem = NULL,
  dem_buffer = 2500,
  max_iterations = 10,
  capping_method = "shortest-path",
  angle_threshold = 100,
  corridor = TRUE,
  segments = FALSE,
  riverspace = FALSE,
  force_download = FALSE,
  ...
)

Arguments

city_name

A place name as a string

river_name

A river name as a string

crs

The projected Coordinate Reference System (CRS) to use. If not provided, the suitable Universal Transverse Mercator (UTM) CRS is selected

network_buffer

Add a buffer (an integer in meters) around river to retrieve additional data (streets, railways, etc.). Default is 3000 m.

buildings_buffer

Add a buffer (an integer in meters) around the river to retrieve additional data (buildings). Default is 100 m.

corridor_init

How to estimate the initial guess of the river corridor. It can take the following values:

"valley": use the river valley boundary, as estimated from a Digital Elevation Model (DEM) (for more info see delineate_valley())

numeric or integer: use a buffer region of the given size (in meters) around the river centerline

An sf::sf or sf::sfc object: use the given input geometry

dem

Digital elevation model (DEM) of the region (only used if corridor_init is "valley")

dem_buffer

Size of the buffer region (in meters) around the river to retrieve the DEM (only used if corridor_init is "valley" and dem is NULL).

max_iterations

Maximum number of iterations employed to refine the corridor edges (see corridor_edge()).

capping_method

The method employed to connect the corridor edge end points (i.e. to "cap" the corridor). See cap_corridor() for the available methods

angle_threshold

Only network edges forming angles above this threshold (in degrees) are considered when forming segment edges. See delineate_segments() and rcoins::stroke(). Only used if segments is TRUE.

corridor

Whether to carry out the corridor delineation

segments

Whether to carry out the corridor segmentation

riverspace

Whether to carry out the riverspace delineation

force_download

Download data even if cached data is available

...

Additional (optional) input arguments for retrieving the DEM dataset (see get_dem()). Only relevant if corridor_init is "valley" and dem is NULL

Value

A list with the corridor, segments, and riverspace geometries

Examples

if (FALSE) { # interactive()
delineate("Bucharest", "Dâmbovița")
}