Identify network edges that are intersecting a geometry

Identify network edges that are intersecting a geometry

Usage

get_intersecting_edges(network, geometry, index = FALSE)

Arguments

network

A spatial network object of class sfnetworks::sfnetwork

geometry

An object of class sf::sfc

index

Whether to return the indices of the matchin edges or the geometries

Value

Indices or geometries of the edges intersecting the given geometry of class sf::sfc_LINESTRING